@@ -1,3 +1,9 @@
 class AdminController < ApplicationController
-  before_action :admin_user
+  before_action :authenticate_admin
+
+  def become
+    user = User.find(params[:id])
+    sign_in(user, { bypass: true })
+    redirect_to root_url, notice: "Impersonating #{user.email}"
+  end
 end
