@@ -1,6 +1,6 @@
 class RunsController < ApplicationController
   before_action :load_entry_and_timesheet, except: [:edit, :update, :destroy]
-  before_action :admin_user, only: [:new, :edit, :create, :destroy]
+  before_action :authenticate_admin, only: [:new, :edit, :create, :destroy]
 
   def new
     @run = @entry.runs.new
@@ -11,6 +11,7 @@ class RunsController < ApplicationController
 
   def edit
     @run = Run.find(params[:id])
+    @statuses = Run.statuses
   end
 
   def create
@@ -43,7 +44,7 @@ class RunsController < ApplicationController
   private
 
     def run_params
-      params.require(:run).permit(:entry_id, :start, :split2, :split3, :split4, :split5, :finish)
+      params.require(:run).permit(:entry_id, :start, :split2, :split3, :split4, :split5, :finish, :status)
     end
 
   def load_entry_and_timesheet
