@@ -1,55 +1,8 @@
 class UsersController < ApplicationController
-  before_action :logged_in_user, only: [:index, :edit, :update]
-  before_action :correct_user, only: [:edit, :update]
-  before_action :admin_user, only: :destroy
-  
-  def index
-    @users = User.all
-  end
-
-  def new
-    @user = User.new
-  end
+  before_action :authenticate_user!, only: [:show]
 
   def show
     @user = User.find(params[:id])
   end
-
-  def create
-    @user = User.new(user_params)
-    if @user.save
-      log_in @user
-      flash[:success] = "Welcome to Sledsheet, #{@user.name}!"
-      redirect_to @user
-    else
-      render 'new'
-    end
-  end
-
-  def edit
-    @user = User.find(params[:id]) # defined in before filter correct_user
-  end
-
-  def update
-    if @user.update_attributes(user_params)
-      flash[:success] = "Profile updated."
-      redirect_to @user
-    else
-      render 'edit'
-    end
-  end
-
-  def destroy
-    User.find(params[:id]).destroy
-    flash[:success] = "User deleted."
-    redirect_to users_url
-  end
-
-  private
-
-    def user_params
-      params.require(:user).permit(:name, :email, :password,
-                                   :password_confirmation)
-    end
 
 end
