@@ -6,18 +6,16 @@ class Entry < ActiveRecord::Base
   validates :athlete_id, :timesheet_id, presence: true
 
   acts_as_list :scope => :timesheet, :column => :bib
-  
+
+  enum status: [:ok, :dns, :dnf, :dsq]
+
   scope :medals, -> { where('position <= 3')} # and timesheet.race
   scope :podiums, -> { where ('position <= 6')}
-  
-  def total_time
-    runs.sum(:finish)
-  end
-  
+
   def date
     timesheet.date
   end
-  
+
   def self.top_ten
     select('athlete_id, count(athlete_id)').
     where('bib <= 3'). # obviously needs to be rewritten
@@ -26,5 +24,10 @@ class Entry < ActiveRecord::Base
     limit(10)
     # BOOM. returns array of relations. first.athlete.name, first.count.
   end
-  
+
+# necessary at the moment for PDF generation
+  def total_time
+    runs.sum(:finish)
+  end
+
 end
