@@ -13,11 +13,6 @@ class AthletesControllerTest < ActionController::TestCase
     assert_response :success
     assert_template 'athletes/index'
     assert_select "title", "Sledsheet | Athletes"
-    assert_not_nil assigns(:athletes)
-    assert_not_nil assigns(:top_ten)
-    Athlete.all.each do |athlete|
-      assert_select 'a[href=?]', athlete_path(athlete), text: athlete.name
-    end
   end
 
   test "should show athlete" do
@@ -28,33 +23,33 @@ class AthletesControllerTest < ActionController::TestCase
   end
 
   test "admin should get new" do
-    log_in_as(@admin)
+    sign_in @admin
     get :new
     assert_response :success
     assert_select "title", "Sledsheet | New Athlete"
   end
 
   test "user should not get new" do
-    log_in_as @user
+    sign_in @user
     get :new
     assert_redirected_to root_path
   end
 
   test "admin should get edit" do
-    log_in_as @admin
+    sign_in @admin
     get :edit, id: @athlete
     assert_response :success
     assert_select "title", "Sledsheet | Edit #{@athlete.name}"
   end
 
   test "user should not get edit" do
-    log_in_as @user
+    sign_in @user
     get :edit, id: @athlete
     assert_redirected_to root_url
   end
 
   test "should create athlete" do
-    log_in_as @admin
+    sign_in @admin
     assert_difference('Athlete.count') do
       post :create, athlete: { first_name: 'Morgan', last_name: 'Tracey', country_code: 'US' }
     end
@@ -62,7 +57,7 @@ class AthletesControllerTest < ActionController::TestCase
   end
 
   test "should update athlete" do
-    log_in_as @admin
+    sign_in @admin
     patch :update, id: @athlete, athlete: { first_name: 'John', last_name: 'Tress', country_code: 'US' }
     assert_redirected_to athlete_path(assigns(:athlete))
   end
