@@ -1,50 +1,27 @@
 require 'test_helper'
 
 class UsersLoginTest < ActionDispatch::IntegrationTest
-  
+
   def setup
     @user = users(:kyle)
   end
-  
+
   test "login with invalid information" do
-    get login_path
-    assert_template 'sessions/new'
-    post login_path, session: { email: "", password: "" }
-    assert_template 'sessions/new'
+    get new_user_session_path
+    assert_template 'devise/sessions/new'
+    post user_session_path, session: { email: "", password: "" }
+    assert_template 'devise/sessions/new'
     assert_not flash.empty?
     get root_path
     assert flash.empty?
   end
-  
-  test "login with valid information followed by logout" do
-    get login_path
-    post login_path, session: { email: @user.email, password: 'password' }
-    assert is_logged_in?
-    assert_redirected_to @user
-    follow_redirect!
-    assert_template 'users/show'
-    assert_select "a[href=?]", login_path, count: 0
-    assert_select "a[href=?]", logout_path
-    assert_select "a[href=?]", user_path(@user)
-    delete logout_path
-    assert_not is_logged_in?
-    assert_redirected_to root_url
-    # Simulate a user clicking logout in a second window.
-    delete logout_path
-    follow_redirect!
-    assert_select "a[href=?]", login_path
-    assert_select "a[href=?]", logout_path,      count: 0
-    assert_select "a[href=?]", user_path(@user), count: 0
-  end
-  
-  test "login with remembering" do
-    log_in_as(@user, remember_me: '1')
-    assert_not_nil cookies['remember_token']
+
+  test "login with valid information" do
+    get new_user_session_path
+    assert_template 'devise/sessions/new'
+    post user_session_path, session: { email: @user.email, password: 'password' }
+    #assert_redirected_to @user
+    assert_template 'user/show'
   end
 
-  test "login without remembering" do
-    log_in_as(@user, remember_me: '0')
-    assert_nil cookies['remember_token']
-  end
-  
 end
